requirements.txt
asyncpg==0.26.0
coc.py==2.1.1
requests==2.28.1
oauth2client==4.1.3
gspread==5.4.0
disnake==2.5.2
pandas==1.4.3
